pyproject.toml
[tool.poetry]
name = "telexkcdbot"
version = "1.0.0"
description = "telexkcdbot"
authors = ["coryphoenixxx <coryphoenixxx@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/coryphoenixxx/telexkcdbot"
packages = [
    { include = "shared", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.11"


[tool.poetry.group.api.dependencies]
fastapi = "^0.109"
python-multipart = "^0.0.6"
orjson = "^3.9.10"
sqlalchemy = "^2.0.23"
asyncpg = "^0.29.0"
uvicorn = "^0.27.0"
pydantic-settings = "^2.1.0"
aiofiles = "^23.2.1"
alembic = "^1.13.0"
pillow = "^10.1.0"
uvloop = "^0.19.0"
filetype = "^1.2.0"
imagesize = "^1.4.1"
python-slugify = "^8.0.1"
httptools = "^0.6.1"
aiohttp = "^3.9.1"
aiodns = "^3.1.1"
aiohttp-retry = "^2.8.3"
gunicorn = "^21.2.0"


[tool.poetry.group.scraper.dependencies]
aiohttp = "^3.9"
uvloop = "^0.19.0"
faust-cchardet = "^2.1.19"
aiodns = "^3.0.0"
lxml = "^5.1.0"
beautifulsoup4 = "^4.12.2"
aiohttp-retry = "^2.8.3"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3"
ruff = "^0.1.9"
black = "^23.12.1"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.0"


[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = ["-s", "-v"]
pythonopath = [".venv/bin"]


[tool.ruff]
ignore-init-module-imports = true
select = ["B", "E", "W", "F", "N", "COM", "I", "ASYNC", "PL", "RUF", "SIM", "C90", "ARG", "T20", "C4", "UP"]
ignore = ["N805", "PLR0913", "PLR2004", "B904", "PLR0911", "PLR0912", "C901", "B008", "ARG001", "RUF013"]
fixable = ["A", "B", "C", "D", "E", "F", "G", "I", "N", "Q", "S", "T", "W", "ANN", "ARG", "BLE", "COM", "DJ", "DTZ", "EM", "ERA", "EXE", "FBT", "ICN", "INP", "ISC", "NPY", "PD", "PGH", "PIE", "PL", "PT", "PTH", "PYI", "RET", "RSE", "RUF", "SIM", "SLF", "TCH", "TID", "TRY", "UP", "YTT"]
unfixable = []

show-fixes = true

exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]
per-file-ignores = { }

line-length = 99

dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

target-version = "py311"


[tool.ruff.mccabe]
max-complexity = 10


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
